+++ laktory/models/pipeline/orchestrators/lakeflowdeclarativepipelineorchestrator.py
@@ -141,6 +141,7 @@ class LakeflowDeclarativePipelineOrchestrator(Pipeline, PipelineChild):
         excludes["type"] = True
         excludes["dataframe_backend"] = True
         excludes["dataframe_api"] = True
+        excludes["purge_mode"] = True
         return excludes
 
     @property