+++ laktory/models/pipeline/orchestrators/sparkdeclarativepipelineorchestrator.py
@@ -110,6 +110,7 @@ class SparkDeclarativePipelineOrchestrator(PipelineChild):
                     "type",
                     "dataframe_backend",
                     "dataframe_api",
+                    "purge_mode",
                     "configuration",
                 ],
                 by_alias=True,